pyproject.toml
[tool.poetry]
name = "auto_nmap"
version = "0.1.0"
description = "A tool to automate nmap scripting through python.poetry "
authors = ["Samuel Owen <SamHirael@gmail.com>", "Oliver Megan <OLI-PLEASE-ADD-YOUR-EMAIL-ADDRESS HERE!>"]

[tool.poetry.dependencies]
python = "^3.9"
python-nmap = "^0.7.1"
netifaces = "^0.11.0"
mock = "^4.0.3"
pytest-mock = "^3.8.2"
pytest-console-scripts = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
discoverDevices = "auto_nmap.discoverDevices:main"
discoverPorts = "auto_nmap.discoverPorts:main"